+++ apps/activitypub/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tryghost/activitypub",
-  "version": "1.0.32",
+  "version": "1.0.33",
   "license": "MIT",
   "repository": {
     "type": "git",
