+++ apps/activitypub/src/components/layout/Layout.tsx
@@ -7,6 +7,7 @@ import {Navigate, ScrollRestoration} from '@tryghost/admin-x-framework';
 import {useAppBasePath} from '@src/hooks/use-app-base-path';
 import {useCurrentPage} from '@src/hooks/use-current-page';
 import {useCurrentUser} from '@tryghost/admin-x-framework/api/currentUser';
+import {useFeatureFlags} from '@src/lib/feature-flags';
 import {useKeyboardShortcuts} from '@hooks/use-keyboard-shortcuts';
 
 const Layout: React.FC<React.HTMLAttributes<HTMLDivElement>> = ({children, ...props}) => {
@@ -16,6 +17,7 @@ const Layout: React.FC<React.HTMLAttributes<HTMLDivElement>> = ({children, ...pr
     const containerRef = useRef<HTMLDivElement>(null);
     const [isMobileSidebarOpen, setIsMobileSidebarOpen] = useState(false);
     const currentPage = useCurrentPage();
+    const {isEnabled} = useFeatureFlags();
 
     const {isNewNoteModalOpen, setIsNewNoteModalOpen} = useKeyboardShortcuts();
 
@@ -44,7 +46,7 @@ const Layout: React.FC<React.HTMLAttributes<HTMLDivElement>> = ({children, ...pr
                         <div className='block grid-cols-[auto_320px] items-start lg:grid'>
                             <div className='z-0 min-w-0'>
                                 <Header
-                                    showBorder={!(currentPage === 'reader')}
+                                    showBorder={!(currentPage === 'reader' || (currentPage === 'explore' && isEnabled('explore-topic')))}
                                     onToggleMobileSidebar={toggleMobileSidebar}
                                 />
                                 <div className='px-[min(4vw,32px)]'>
