+++ apps/activitypub/src/lib/feature-flags.tsx
@@ -2,7 +2,7 @@ import React, {createContext, useContext, useEffect, useState} from 'react';
 import {useLocation} from '@tryghost/admin-x-framework';
 
 // Define all available feature flags as string here, e.g. ['flag-1', 'flag-2']
-export const FEATURE_FLAGS = [] as const;
+export const FEATURE_FLAGS = ['explore-topic'] as const;
 
 // ---
 export type FeatureFlag = typeof FEATURE_FLAGS[number] | string;
